Refactor installation screen
